% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMRMA.R
\name{addGene}
\alias{addGene}
\title{add genes}
\usage{
addGene(geneCellTypeList, celltype)
}
\arguments{
\item{geneCellTypeList}{user-provided genes and cell types added to existed reference}

\item{celltype}{existed reference cell types}
}
\value{
cell type with added genes
}
\description{
add genes to existed cell type
}
