% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMRMA.R
\name{databaseVisual}
\alias{databaseVisual}
\title{generate a html file to visualize hierarchical cell type structure}
\usage{
databaseVisual(celltype, repeated = F)
}
\arguments{
\item{celltype}{reference cell type for visualization}
}
\value{
html file
}
\description{
generate a html file to visualize hierarchical cell type structure
}
