% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMRMA.R
\name{get_celltype}
\alias{get_celltype}
\title{get hierarchical database}
\usage{
get_celltype(species, db)
}
\arguments{
\item{species}{choose species}

\item{db}{hierarchical database}
}
\value{
CellType
}
\description{
get hierarchical database
}
