% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMRMA.R
\name{get_variableGene}
\alias{get_variableGene}
\title{get_variableGene}
\usage{
get_variableGene(data, chunk = 1000)
}
\arguments{
\item{data}{normalized data}
}
\value{
hvf.info
}
\description{
find variable genes based on normalized data
}
