% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogNorm}
\alias{LogNorm}
\title{normalize single cell count data}
\usage{
LogNorm(data, display_progress = FALSE)
}
\arguments{
\item{data}{count data}
}
\description{
normalize single cell count data
}
