% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMRMA.R
\name{ORA_celltype}
\alias{ORA_celltype}
\title{ORA celltype}
\usage{
ORA_celltype(weight, meanexp, cellType, rough)
}
\arguments{
\item{weight}{weight for each gene}

\item{meanexp}{mean normalized data for each cluster}

\item{cellType}{reference cell type}

\item{rough}{rough=T when level 1}
}
\value{
list
}
\description{
detailed annotation algorithm
}
