% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMRMA.R
\name{ori_annotation}
\alias{ori_annotation}
\title{ori_annotation}
\usage{
ori_annotation(marker, clusters, i, sub, p)
}
\arguments{
\item{marker}{reference cell type}

\item{clusters}{cluster information}

\item{i}{the number of level + 1}

\item{sub}{normalized data}

\item{p}{p value cut off}
}
\value{
newcluters cluster with annotaiton information
}
\description{
get annotation process for each cluster
}
