% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMRMA.R
\name{precise_cluster}
\alias{precise_cluster}
\title{precise cluster}
\usage{
precise_cluster(rawdata, logdata, npcs, nhvg, k)
}
\arguments{
\item{npcs}{number of princple components}

\item{scdata}{single cell count data}

\item{logcounts}{normalized data}
}
\value{
clusters
}
\description{
clustering process after level 1
}
