% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMRMA.R
\name{remove_singleton}
\alias{remove_singleton}
\title{remove singleton}
\usage{
remove_singleton(clusters)
}
\arguments{
\item{clusters}{clusters}
}
\value{
clusters
}
\description{
remove clusters with only one cell
}
