% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMRMA.R
\name{rough_cluster}
\alias{rough_cluster}
\title{rough cluster}
\usage{
rough_cluster(scdata, logcounts, nhvg, k)
}
\arguments{
\item{scdata}{single cell count data}

\item{logcounts}{normalized data}
}
\value{
clusters
}
\description{
clustering process for level 1
}
