% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RunModularityClusteringCpp}
\alias{RunModularityClusteringCpp}
\title{lovain modularity}
\usage{
RunModularityClusteringCpp(
  SNN,
  modularityFunction,
  resolution,
  algorithm,
  nRandomStarts,
  nIterations,
  randomSeed,
  printOutput
)
}
\description{
lovain modularity
}
