% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMRMA.R
\name{scMRMA}
\alias{scMRMA}
\title{scMRMA}
\usage{
scMRMA(
  input,
  species,
  db = "panglaodb",
  p = 0.05,
  normalizedData = F,
  selfDB = NULL,
  selfClusters = NULL
)
}
\arguments{
\item{selfClusters}{}
}
\value{
Result
}
\description{
scMRMA single cell annotation

scMRMA single cell annotation
}
\examples{
library(scMRMA)
load(system.file("data", "colon1.Rdata", package = "scMRMA"))
result<-scMRMA(input=colon1, species="Mm")
result
}
