% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMRMA.R
\name{selfDefinedDatabase}
\alias{selfDefinedDatabase}
\title{self defined database}
\usage{
selfDefinedDatabase(file)
}
\arguments{
\item{file}{user-provided reference database}
}
\value{
celltype
}
\description{
user-provided hierarchical cell type
}
