% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMRMA.R
\name{uniformR_anno}
\alias{uniformR_anno}
\title{uniform resolution annotation}
\usage{
uniformR_anno(CellType, scdata, logcounts, p, selfClusters, nhvg, k)
}
\arguments{
\item{CellType}{reference cell type}

\item{scdata}{single cell count data}

\item{logcounts}{single cell normalized data}

\item{p}{p value cutoff}

\item{selfClusters}{user-provided cluster information}
}
\value{
uniformR
}
\description{
uniform resolution annotation
}
